requirements.txt
discord.py
tasks